+++ test/run_test.py
@@ -39,6 +39,7 @@ TESTS = [
     'nn',
     'numba_integration',
     'optim',
+    'quantized',
     'sparse',
     'thd_distributed',
     'torch',
